+++ ultralytics/utils/export/engine.py
@@ -3,12 +3,43 @@
 from __future__ import annotations
 
 import json
+import types
 from pathlib import Path
 
 import torch
 
-from ultralytics.utils import IS_JETSON, LOGGER
-from ultralytics.utils.torch_utils import TORCH_2_4
+from ultralytics.utils import IS_JETSON, LOGGER, TORCH_VERSION
+from ultralytics.utils.torch_utils import TORCH_2_4, TORCH_2_9
+
+
+def best_onnx_opset(onnx: types.ModuleType, cuda: bool = False) -> int:
+    """Return max ONNX opset for this torch version with ONNX fallback."""
+    if TORCH_2_4:  # _constants.ONNX_MAX_OPSET first defined in torch 1.13
+        opset = torch.onnx.utils._constants.ONNX_MAX_OPSET - 1  # use second-latest version for safety
+        if TORCH_2_9:
+            opset = min(opset, 20)  # legacy TorchScript exporter caps at opset 20 in torch 2.9+
+        if cuda:
+            opset -= 2  # fix CUDA ONNXRuntime NMS squeeze op errors
+    else:
+        version = ".".join(TORCH_VERSION.split(".")[:2])
+        opset = {
+            "1.8": 12,
+            "1.9": 12,
+            "1.10": 13,
+            "1.11": 14,
+            "1.12": 15,
+            "1.13": 17,
+            "2.0": 17,  # reduced from 18 to fix ONNX errors
+            "2.1": 17,  # reduced from 19
+            "2.2": 17,  # reduced from 19
+            "2.3": 17,  # reduced from 19
+            "2.4": 20,
+            "2.5": 20,
+            "2.6": 20,
+            "2.7": 20,
+            "2.8": 23,
+        }.get(version, 12)
+    return min(opset, onnx.defs.onnx_opset_version())
 
 
 def torch2onnx(
